# Use unbuffered I/O for faster printing to stdout, switch to Rustls