--- a/Cargo.toml
+++ b/Cargo.toml
@@ -23,7 +23,7 @@ docopt = "1"
 env_logger = { version = "0.7", optional = true }
 flate2 = "1"
 log = "0.4"
-reqwest = { version = "0.10.3", features = ["blocking", "native-tls"], default-features = false }
+reqwest = { version = "0.10.3", features = ["blocking", "rustls-tls"], default-features = false }
 serde = "1.0.21"
 serde_derive = "1.0.21"
 tar = "0.4.14"