- Use unbuffered IO for faster printing, see https://llogiq.github.io/2017/06/01/perf-pitfalls.html for more
- Use statically linked `rustls` over dynamically linked `openssl`.

Compared to the benchmark in #129 from the first post (ran `hyperfine --prepare 'sync; echo 3 | sudo tee /proc/sys/vm/drop_caches' './target/release/tldr tar'`). I don't know the exact command that the author of `outfieldr` ran, but you can also see a comparison I posted there. 

```
Benchmark #1: ./target/release/tldr tar
  Time (mean ± σ):      29.3 ms ±   6.2 ms    [User: 1.7 ms, System: 4.3 ms]
  Range (min … max):    24.3 ms …  50.1 ms    22 runs
```