@sondr3 2021-05-13
Updated according to the reviews, I'm on a different computer so can't run the benchmark again but there shouldn't be a difference.