@sondr3 2021-05-16
I actually started converting the `unwrap`s to proper errors, but figured it was out of the scope of this PR. Happy to fix though, it may take a few days as we are celebrating our Independence Day tomorrow 😄